% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_BoP_to_analysis_intervals.R
\name{attach_BoP_to_analysis_intervals}
\alias{attach_BoP_to_analysis_intervals}
\title{attach BoP to interval built with determine_analysis_intervals function}
\usage{
attach_BoP_to_analysis_intervals(all_db_data, analysis_intervals,
  subsequenceSize = 4 * 60, alphabetSize = 5, wordSize = 4,
  doNumerosityReduction = TRUE)
}
