% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_data_by_interval_type2.R
\name{cut_data_by_interval_type2}
\alias{cut_data_by_interval_type2}
\title{Put all data into a dataframe, labelled by the kind of interval they are}
\usage{
cut_data_by_interval_type2(annotation_data, heart_beat_data)
}

