% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_analysis_intervals.R
\name{determine_analysis_intervals}
\alias{determine_analysis_intervals}
\title{funkcija ki vrne vse interval dolo훾enega tipa in pri kateri to훾ki naj odre탑e podatke}
\usage{
determine_analysis_intervals(all_db_data_info, Interval_type,
  Min_interval_length, Keep_total_interval_length = NULL,
  Cut_off_last_X_seconds)
}
\description{
to je input za funkcijo, ki ra훾una BoPe
}
