% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_intervals_for_all_records_in_database.R
\name{extract_intervals_for_all_records_in_database}
\alias{extract_intervals_for_all_records_in_database}
\title{extract intervals from a database}
\usage{
extract_intervals_for_all_records_in_database(database)
}

