% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_annotation.R
\name{find_annotation}
\alias{find_annotation}
\title{Find annotations in annotation txt file}
\usage{
find_annotation(annotation_time = "Seconds", annotation_data,
  lookup_column = "Aux", lookup_texts)
}
