% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportHRConstantInterval.R
\name{ImportHRConstantInterval}
\alias{ImportHRConstantInterval}
\title{Import physionet generated data, saved in a txt file}
\usage{
ImportHRConstantInterval(workingDir, fileSuffix = ".unauditedHRconstint.txt")
}
\arguments{
\item{workingDir}{The directory the data and RECORDS file (containing
the list of samples) are located.}

\item{fileSuffix}{The suffix of the files from which data should be imported.
E.g. "".unauditedHRconstint.txt" by default.}
}
\description{
Imports data from a txt file (csv), that contains two columns:
time of measurement in seconds and
heart rate in beats per minute (BPM). 
Meant to be used on constant interval measurements.
Outputs a data frame wih same two cloumns.
}

