% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_lines_with_pattern.R
\name{remove_lines_with_pattern}
\alias{remove_lines_with_pattern}
\title{Remove lines with some pattern}
\usage{
remove_lines_with_pattern(file, pattern)
}
\arguments{
\item{file}{file to check}

\item{pattern}{pattern in line to remove}
}

