% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_interval_boundaries.R
\name{return_interval_boundaries}
\alias{return_interval_boundaries}
\title{Returns data frame of interval boundaries}
\usage{
return_interval_boundaries(annotation_data, heart_beat_data,
  event_start_signal = "(AFIB", event_end_signal = "(N")
}
