% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_physionet_HR.R
\name{Import_physionet_HR}
\alias{Import_physionet_HR}
\title{Import physionet generated data, saved in a txt file

# TODO : redo help section after making the function more general}
\usage{
Import_physionet_HR(workingDir, fileSuffix_HR_constint, fileSuffix_annotations)
}
\arguments{
\item{workingDir}{The directory the data and RECORDS file (containing
the list of samples) are located.}

\item{fileSuffix_HR_constint}{The suffix of the files from which HR constant interval data should be imported.
E.g. "".unauditedHRconstint.txt" by default.}

\item{fileSuffix_annotations}{Suffix for annotations file.}
}
\description{
Imports data from a txt file (csv), that contains two columns:
time of measurement in seconds and
heart rate in beats per minute (BPM). 
Meant to be used on constant interval measurements.
Outputs a data frame wih same two cloumns.
}
